+++ functions/src/members/createMemberSearchIndex.ts
@@ -1,6 +1,7 @@
 import { pubsub } from "firebase-functions"
 import { db } from "../firebase"
 import * as api from "../malegislature"
+import { currentGeneralCourt } from "../../../components/db/common"
 
 /** Create a document that aggregates all legislative members for easier
  * searching on the client.  */
@@ -8,7 +9,7 @@ export const createMemberSearchIndex = pubsub
   .schedule("every 24 hours")
   .onRun(async () => {
     const members = await db
-      .collection(`/generalCourts/${api.currentGeneralCourt}/members`)
+      .collection(`/generalCourts/${currentGeneralCourt}/members`)
       .get()
 
     const index = members.docs
@@ -19,7 +20,7 @@ export const createMemberSearchIndex = pubsub
       .sort((m1, m2) => (m1.Name < m2.Name ? -1 : 1))
 
     await db
-      .doc(`/generalCourts/${api.currentGeneralCourt}/indexes/memberSearch`)
+      .doc(`/generalCourts/${currentGeneralCourt}/indexes/memberSearch`)
       .set({
         representatives: index.filter(d => d.Branch === "House"),
         senators: index.filter(d => d.Branch === "Senate")
